Add lots more fractals, modularise
